FourierTransform: 	- added freqToIndex method for converting a frequency to an index (duh) 	- added getAverageCenterFrequency to get the center frequency of an average band from an index 	- added forward(float[], int) and forward(AudioBuffer,int) so an offset into the array of floats can be 	  provided, in case a user has a large buffer they want to analyze only parts of.
JSMinim:
    - added setInputMixer(Mixer) so users can specify which Mixer to use when getting TargetDataLines
    - added setOutputMixer(Mixer) so users can specify which Mixer to use when getting SourceDataLines
    
JSMPEGAudioRecordingStream:
	- removed TODO about using skip because using skip doesn't work well.
	
JSPCMAudioRecordingStream:
	- removed TODO about using skip because using skip doesn't work well.
	
Minim:
	- added getServiceProvider() so that you can get a handle on the implementation being used by Minim.
	- added createSample functions to get an AudioSample using a buffer(s) of floats.
	
MinimServiceProvider:
	- added getAudioSample(float[], AudioFormat, int) and getAudioSample(float[], float[], AudioFormat, int)
	  interface functions for requesting AudioSamples with buffers instead of from files.

+++ src/ddf/minim/javasound/JSPCMAudioRecordingStream.java
@@ -79,10 +79,15 @@ class JSPCMAudioRecordingStream extends JSBaseAudioRecordingStream
 		{
 			while (totalSkipped < toSkip)
 			{
-				int read;
+				long read;
 				synchronized ( ais )
 				{
-          // TODO: why aren't we using skip here?
+          // we don't use skip here because it sometimes has problems where
+          // it's "unable to skip an integer number of frames",
+          // which sometimes means it doesn't skip at all and other times
+          // means that you wind up with noise because it lands at half
+          // a sample off from where it should be. 
+          // read seems to be rock solid.
 					read = ais.read(skipBytes, 0, (int)(toSkip - totalSkipped));
 				}
 				if (read == -1)
